wip: move services to shared folder
