+++ source/databricks/geh_stream/shared/services/__init__.py
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from .coordinator_service import CoordinatorService
-from .aggregation_postprocessor import PostProcessor
+from .postprocessor import PostProcessor
